+++ SuperDUPER/BasicSword.cpp
@@ -9,8 +9,8 @@ BasicSword::BasicSword()
 
 void BasicSword::render(SDL_Renderer* renderer, const Camera& camera, const LifeForm& owner) const {
 	if (camera.isInCamera(Rectangle{ static_cast<int>(owner.getPosition().x), static_cast<int>(owner.getPosition().y), 1, 1 })) {
-		Position<> front{ camera.absoluteToRelative(owner.getPosition().x + owner.getRadius(), owner.getPosition().y + owner.getRadius()) };
-		Position<> back{ camera.absoluteToRelative(owner.getPosition().x - owner.getRadius(), owner.getPosition().y - owner.getRadius()) };
+		Position<> front{ camera.absoluteToRelative(static_cast<long>(owner.getPosition().x) + owner.getRadius(), static_cast<long>(owner.getPosition().y) + owner.getRadius()) };
+		Position<> back{ camera.absoluteToRelative(static_cast<long>(owner.getPosition().x) - owner.getRadius(), static_cast<long>(owner.getPosition().y) - owner.getRadius()) };
 		SDL_SetRenderDrawColor(renderer, 128, 128, 255, 255);
 		SDL_RenderDrawLine(renderer, front.x, front.y, back.x, back.y);
 	}
